LLM/requirements.txt
llama-cpp-python==0.2.20
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.4.2
numpy==1.24.3